Use upstream CU masking HIP API
 - This change removes the dependency on my modified version of HIP in
   order to set CU masks for streams, due to
   hipExtStreamCreateWithCUMask being added to ROCm 3.6.

 - hipExtStreamCreateWithCUMask expects the CU mask to be a list of
   32-bit values, so multiple files were changed to use 32-bit rather
   than 64-bit CU masks.

 - Added a couple configs for testing performance of large memory
   transfers being subdivided into smaller chunks.

 - Fixed a python 2 vs. 3 compatibility issue in view_times_cdf.py.

 - Updated configs/counter_spi_half_vs_full_streams.json to use 60 vs.
   30 CUs on the Radeon VII.

+++ scripts/view_times_cdf.py
@@ -68,7 +68,7 @@ def get_plugin_cdf(plugin, times_key):
         if not times_key in t:
             continue
         times = t[times_key]
-        for i in range(len(times) / 2):
+        for i in range(int(len(times) / 2)):
             start_index = i * 2
             end_index = i * 2 + 1
             raw_values.append(times[end_index] - times[start_index])
